requirements.txt
transformers>=4.36.0
groq>=0.4.0
asyncio>=3.4.3
aiohttp>=3.9.0
pandas>=2.1.0
pytest>=7.4.0
python-dotenv>=1.0.0
